requirements.txt
# Core Dependencies
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=0.0.5
langchain-community>=0.0.20
google-generativeai>=0.8.0

# Vector Store
faiss-cpu>=1.7.4

# API
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Data Science
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# ML Lifecycle
mlflow>=2.10.0

# Document Processing
PyPDF2>=3.0.0

# Environment
python-dotenv>=1.0.0

